Initial move of datalayer plugins to epp/framework/plugins
