+++ pkg/epp/framework/plugins/datalayer/attribute/prefix/data_types.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package approximateprefix
+package prefix
 
 import (
 	fwkdl "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/framework/interface/datalayer"
