+++ pkg/epp/framework/plugins/datalayer/source/metrics/factories.go
@@ -27,8 +27,8 @@ import (
 	"github.com/prometheus/common/model"
 	flag "github.com/spf13/pflag"
 
-	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/datalayer/http"
 	fwkplugin "sigs.k8s.io/gateway-api-inference-extension/pkg/epp/framework/interface/plugin"
+	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/framework/plugins/datalayer/source/http"
 )
 
 const (
